v0.10.7.6 commit

+++ NBA Stats Tracker/Data/NST.Data.Players.cs
@@ -920,7 +920,7 @@ namespace NBA_Stats_Tracker.Data
                         break;
 
                     case "ftrn":
-                        s += String.Format("FTM/FGA: {0}-{1} ({2:F3})\n", FTM, FGA, FTM/FGA);
+                        s += String.Format("FTM/FGA: {0}-{1} ({2:F3})\n", FTM, FGA, (double)FTM/FGA);
                         break;
                 }
